feat(Josephus): create Josephus module
