+++ src/components/Josephus/CircleArea/index.ts
@@ -1,0 +1,1 @@
+export { default as CircleArea } from './CircleArea';
