+++ src/components/Josephus/KillerInput/index.ts
@@ -1,0 +1,1 @@
+export { default as KillerInput } from './KillerInput';
