+++ src/components/Josephus/People/People.tsx
