+++ src/components/Josephus/People/index.ts
@@ -1,0 +1,1 @@
+export { default as People } from './People';
