+++ src/components/Josephus/hooks/useDeadSequence.ts
