+++ src/components/Josephus/index.ts
@@ -1,0 +1,1 @@
+export { default as Josephus } from './Josephus';
